Remove unnecessary boxing and unboxing

+++ src/main/java/org/codehaus/groovy/runtime/typehandling/IntegerMath.java
@@ -30,19 +30,19 @@ public final class IntegerMath extends NumberMath {
     private IntegerMath() {}
                     
     protected Number absImpl(Number number) {
-        return Integer.valueOf(Math.abs(number.intValue()));
+        return Math.abs(number.intValue());
     }
     
     public Number addImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() + right.intValue());
+        return left.intValue() + right.intValue();
     }
 
     public Number subtractImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() - right.intValue());
+        return left.intValue() - right.intValue();
     }
 
     public Number multiplyImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() * right.intValue());
+        return left.intValue() * right.intValue();
     }
 
     public Number divideImpl(Number left, Number right) {
@@ -56,46 +56,46 @@ public final class IntegerMath extends NumberMath {
     }
 
     protected Number orImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() | right.intValue());
+        return left.intValue() | right.intValue();
     }
 
     protected Number andImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() & right.intValue());
+        return left.intValue() & right.intValue();
     }
 
     protected Number xorImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() ^ right.intValue());
+        return left.intValue() ^ right.intValue();
     }
 
     protected Number intdivImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() / right.intValue());
+        return left.intValue() / right.intValue();
     }
     
     protected Number modImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() % right.intValue());
+        return left.intValue() % right.intValue();
     }
 
     protected Number unaryMinusImpl(Number left) {
-        return Integer.valueOf(-left.intValue());
+        return -left.intValue();
     }
 
     protected Number unaryPlusImpl(Number left) {
-        return Integer.valueOf(left.intValue());
+        return left.intValue();
     }
 
     protected Number bitwiseNegateImpl(Number left) {
-        return Integer.valueOf(~left.intValue());
+        return ~left.intValue();
     }
 
     protected Number leftShiftImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() << right.intValue());
+        return left.intValue() << right.intValue();
     }
 
     protected Number rightShiftImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() >> right.intValue());
+        return left.intValue() >> right.intValue();
     }
 
     protected Number rightShiftUnsignedImpl(Number left, Number right) {
-        return Integer.valueOf(left.intValue() >>> right.intValue());
+        return left.intValue() >>> right.intValue();
     }
 }
