+++ src/main/java/org/codehaus/groovy/runtime/typehandling/LongMath.java
@@ -30,19 +30,19 @@ public final class LongMath extends NumberMath {
     private LongMath() {}
                     
     protected Number absImpl(Number number) {
-        return Long.valueOf(Math.abs(number.longValue()));
+        return Math.abs(number.longValue());
     }
     
     public Number addImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() + right.longValue());
+        return left.longValue() + right.longValue();
     }
 
     public Number subtractImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() - right.longValue());
+        return left.longValue() - right.longValue();
     }
 
     public Number multiplyImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() * right.longValue());
+        return left.longValue() * right.longValue();
     }
 
     public Number divideImpl(Number left, Number right) {
@@ -56,50 +56,50 @@ public final class LongMath extends NumberMath {
     }
 
     protected Number intdivImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() / right.longValue());
+        return left.longValue() / right.longValue();
     }
     
     protected Number modImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() % right.longValue());
+        return left.longValue() % right.longValue();
     }
     
     protected Number unaryMinusImpl(Number left) {
-        return Long.valueOf(-left.longValue());
+        return -left.longValue();
     }
 
     protected Number unaryPlusImpl(Number left) {
-        return Long.valueOf(left.longValue());
+        return left.longValue();
     }
 
     protected Number bitwiseNegateImpl(Number left) {
-        return Long.valueOf(~left.longValue());
+        return ~left.longValue();
     }
     
     protected Number orImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() | right.longValue());
+        return left.longValue() | right.longValue();
     }
 
     protected Number andImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() & right.longValue());
+        return left.longValue() & right.longValue();
     }
     
     protected Number xorImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() ^ right.longValue());
+        return left.longValue() ^ right.longValue();
     }
     
     protected Number leftShiftImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() << right.longValue());
+        return left.longValue() << right.longValue();
     }
 
     protected Number rightShiftImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() >> right.longValue());
+        return left.longValue() >> right.longValue();
     }
 
     protected Number rightShiftUnsignedImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() >>> right.longValue());
+        return left.longValue() >>> right.longValue();
     }
 
     protected Number bitAndImpl(Number left, Number right) {
-        return Long.valueOf(left.longValue() & right.longValue());
+        return left.longValue() & right.longValue();
     }
 }
